refactor: consolidate ModelFactory into agents-core
- Move ModelFactory from three duplicate locations (agents-run-api, agents-eval-api, agents-manage-ui) into agents-core package
- Rename file to use kebab-case convention (model-factory.ts)
- Add AI SDK provider dependencies to agents-core (anthropic, openai, google, openrouter, gateway, openai-compatible)
- Update all imports across packages to use @inkeep/agents-core
- Remove duplicate ModelSettings interface, import from validation/schemas instead
- Delete duplicate ModelFactory files from all packages

Benefits:
- Eliminates ~800 lines of duplicate code
- Single source of truth for AI model creation
- Easier maintenance and future updates
- Consistent behavior across all packages
