+++ agents-run-api/src/__tests__/agents/Agent.test.ts
@@ -138,12 +138,14 @@ vi.mock('@ai-sdk/anthropic', () => ({
 }));
 
 // Mock ModelFactory
-vi.mock('../../agents/ModelFactory.js', () => {
+vi.mock('@inkeep/agents-core', async () => {
+  const actual = await vi.importActual('@inkeep/agents-core');
   const mockModel = 'mocked-language-model';
   const mockGenerationParams = { temperature: 0.7, maxTokens: 4096 };
   const mockGenerationConfig = { model: mockModel, ...mockGenerationParams };
 
   return {
+    ...actual,
     ModelFactory: {
       createModel: vi.fn().mockReturnValue(mockModel),
       getGenerationParams: vi.fn().mockReturnValue(mockGenerationParams),
@@ -1257,7 +1259,7 @@ describe('Agent Model Settings', () => {
     await agent.generate('Test prompt');
 
     // Get the mocked ModelFactory
-    const { ModelFactory } = await import('../../agents/ModelFactory.js');
+    const { ModelFactory } = await import('@inkeep/agents-core');
     expect(ModelFactory.prepareGenerationConfig).toHaveBeenCalledWith({
       model: 'anthropic/claude-sonnet-4-5',
       providerOptions: undefined,
@@ -1283,7 +1285,7 @@ describe('Agent Model Settings', () => {
     const agent = new Agent(configWithModel);
     await agent.generate('Test prompt');
 
-    const { ModelFactory } = await import('../../agents/ModelFactory.js');
+    const { ModelFactory } = await import('@inkeep/agents-core');
     expect(ModelFactory.prepareGenerationConfig).toHaveBeenCalledWith({
       model: 'openai/gpt-4o',
       providerOptions: {
@@ -1314,7 +1316,7 @@ describe('Agent Model Settings', () => {
     const agent = new Agent(configWithModel);
     await agent.generate('Test prompt');
 
-    const { ModelFactory } = await import('../../agents/ModelFactory.js');
+    const { ModelFactory } = await import('@inkeep/agents-core');
     expect(ModelFactory.prepareGenerationConfig).toHaveBeenCalledWith(
       expect.objectContaining({
         model: 'anthropic/claude-sonnet-4-5',
@@ -1372,7 +1374,7 @@ describe('Agent Model Settings', () => {
     const agent = new Agent(configWithDataComponents);
     await agent.generate('Test prompt');
 
-    const { ModelFactory } = await import('../../agents/ModelFactory.js');
+    const { ModelFactory } = await import('@inkeep/agents-core');
     // Called twice: once for text generation with custom model, once for structured output with OpenAI model
     expect(ModelFactory.prepareGenerationConfig).toHaveBeenCalledTimes(2);
     expect(ModelFactory.prepareGenerationConfig).toHaveBeenNthCalledWith(1, {
@@ -1404,7 +1406,7 @@ describe('Agent Model Settings', () => {
     const agent = new Agent(configWithDataComponents);
     await agent.generate('Test prompt');
 
-    const { ModelFactory } = await import('../../agents/ModelFactory.js');
+    const { ModelFactory } = await import('@inkeep/agents-core');
     // Called twice: once for text generation, once for structured output (both use base model)
     expect(ModelFactory.prepareGenerationConfig).toHaveBeenCalledTimes(2);
     expect(ModelFactory.prepareGenerationConfig).toHaveBeenNthCalledWith(1, {
@@ -1430,7 +1432,7 @@ describe('Agent Model Settings', () => {
     const agent = new Agent(configWithOpenAI);
     await agent.generate('Test prompt');
 
-    const { ModelFactory } = await import('../../agents/ModelFactory.js');
+    const { ModelFactory } = await import('@inkeep/agents-core');
     expect(ModelFactory.prepareGenerationConfig).toHaveBeenCalledWith({
       model: 'openai/gpt-4o',
       providerOptions: undefined,
@@ -1450,7 +1452,7 @@ describe('Agent Model Settings', () => {
     const agent = new Agent(configWithPlainModel);
     await agent.generate('Test prompt');
 
-    const { ModelFactory } = await import('../../agents/ModelFactory.js');
+    const { ModelFactory } = await import('@inkeep/agents-core');
     expect(ModelFactory.prepareGenerationConfig).toHaveBeenCalledWith({
       model: 'anthropic/claude-3-5-haiku-latest',
       providerOptions: undefined,
