+++ packages/agents-core/package.json
@@ -43,6 +43,11 @@
   },
   "clean-package": "./clean-package.config.json",
   "dependencies": {
+    "@ai-sdk/anthropic": "2.0.2",
+    "@ai-sdk/gateway": "^1.0.23",
+    "@ai-sdk/google": "^2.0.14",
+    "@ai-sdk/openai": "2.0.11",
+    "@ai-sdk/openai-compatible": "^1.0.27",
     "@composio/core": "^0.2.4",
     "@electric-sql/pglite": "^0.3.13",
     "@hono/node-server": "^1.14.3",
@@ -50,6 +55,7 @@
     "@modelcontextprotocol/sdk": "^1.17.2",
     "@nangohq/node": "^0.69.5",
     "@nangohq/types": "^0.69.5",
+    "@openrouter/ai-sdk-provider": "^1.2.0",
     "@opentelemetry/api": "^1.9.0",
     "@opentelemetry/auto-instrumentations-node": "^0.62.0",
     "@opentelemetry/baggage-span-processor": "^0.4.0",
