+++ packages/agents-core/src/utils/index.ts
@@ -8,6 +8,7 @@ export * from './execution';
 export * from './format-messages';
 export * from './logger';
 export * from './mcp-client';
+export * from './model-factory';
 export * from './schema-conversion';
 export * from './service-token-auth';
 export * from './third-party-mcp-servers';
